+++ NEWS.md
@@ -8,6 +8,13 @@
 ## Coercion
 
 - Improved `as_stocnet.array()` to join `from_times()`, `from_layers()`, or `from_reporters()` slices
+- Added `...` to the `as_matrix()` generic, so that methods can take arguments
+- Added conversion from node_measure and node_member to matrices and stocnets (closes #161)
+  - Added `as_matrix.node_measure()` and `as_matrix.node_member()`
+  - Added `as_stocnet.node_measure()` and `as_stocnet.node_member()`
+  - `compare=` takes "absdiff", "diff", "sender", or "receiver" for a measure
+  - `compare=` takes "same" for a membership
+  - `twomode=TRUE` turns a membership into an affiliation matrix
 
 ## Modifying